requirements.txt
playwright==1.54.0
polars==1.32.3
protobuf==6.32.0
python_dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.0
configparser==7.2.0
